requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altgraph==0.17.3
asgiref==3.6.0
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
base58==2.1.1
beautifulsoup4==4.11.2
bitarray==2.7.2
blis==0.7.9
books==0.1.2
bs4==0.0.1
cachetools==5.3.0
catalogue==2.0.8
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
charset-normalizer==2.1.1
ChatterBot==0.8.7
chatterbot-corpus==1.1.4
click==8.1.3
colorama==0.4.6
comtypes==1.1.14
confection==0.0.4
contourpy==1.0.7
crispy-bootstrap5==0.7
cycler==0.11.0
cymem==2.0.7
cytoolz==0.12.1
decorator==5.1.1
distlib==0.3.6
dj-database-url==1.2.0
Django==4.1.6
django-ckeditor==6.5.1
django-cors-headers==3.13.0
django-js-asset==2.0.0
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.5.1
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
filelock==3.9.0
Flask==2.2.2
fonttools==4.38.0
frozenlist==1.3.3
future==0.18.2
geocoder==1.38.1
geographiclib==2.0
geopy==2.3.0
google-api-core==2.11.0
google-api-python-client==2.78.0
google-auth==2.16.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.58.0
h11==0.14.0
hashable==0.1
hexbytes==0.3.0
html5lib==1.1
httplib2==0.21.0
idna==3.4
ipfshttpclient==0.8.0a2
itsdangerous==2.1.2
jaraco.context==4.3.0
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
kiwisolver==1.4.4
langcodes==3.3.0
lru-dict==1.1.8
MarkupSafe==2.1.2
material==0.1
mathparse==0.1.2
matplotlib==3.6.3
more-itertools==9.0.0
MouseInfo==0.1.3
multiaddr==0.0.9
multidict==6.0.4
murmurhash==1.0.9
mysqlclient==2.1.1
netaddr==0.8.0
nltk==3.8.1
numpy==1.24.1
oauthlib==3.2.2
opencv-python==4.7.0.68
outcome==1.2.0
packaging==23.0
pandas==1.5.3
parsimonious==0.8.1
pathy==0.10.1
pefile==2022.5.30
Pillow==9.3.0
Pint==0.20.1
platformdirs==2.6.2
preshed==3.0.8
protobuf==3.19.5
psutil==5.9.4
psycopg2==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.13
PyAutoGUI==0.9.53
pycparser==2.21
pycryptodome==3.17
pydantic==1.10.4
pyfiglet==0.8.post1
PyGetWindow==0.0.9
pyinstaller==5.7.0
pyinstaller-hooks-contrib==2022.14
pyjokes==0.6.0
pymongo==3.13.0
PyMsgBox==1.0.9
pyparsing==3.0.9
pyperclip==1.8.2
pypiwin32==223
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.1
PyRect==0.2.0
pyrsistent==0.19.3
PyScreeze==0.1.28
PySocks==1.7.1
python-dateutil==2.6.1
python-twitter==3.5
pyttsx3==2.90
pytweening==1.0.4
pytz==2022.7.1
pywhatkit==5.4
pywin32==305
pywin32-ctypes==0.2.0
PyYAML==3.13
ratelim==0.1.6
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
rlp==2.0.1
rsa==4.9
scipy==1.10.0
selenium==4.8.2
six==1.16.0
sketchpy==0.1.0
smart-open==6.3.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
spacy==3.5.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SpeechRecognition==3.9.0
SQLAlchemy==1.2.19
sqlparse==0.4.3
srsly==2.4.5
svg.path==6.2
svgpathtools==1.5.1
svgwrite==1.4.3
thinc==8.1.7
toolz==0.12.0
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
turtle==0.0.1
typer==0.7.0
typing_extensions==4.4.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.13
varint==1.0.2
virtualenv==20.17.1
wasabi==1.1.1
web3==5.31.3
webencodings==0.5.1
websockets==9.1
Werkzeug==2.2.2
wikipedia==1.4.0
wolframalpha==5.0.0
wsproto==1.2.0
xmltodict==0.13.0
yarl==1.8.2
